+++ src/ska_low_mccs_pasd/fncc/fncc_device.py
@@ -96,8 +96,7 @@ class MccsFNCC(MccsBaseDevice[FnccComponentManager]):
 
     def delete_device(self: MccsFNCC) -> None:
         """Delete the device."""
-        self.component_manager._pasd_bus_proxy.cleanup()
-        self.component_manager._task_executor._executor.shutdown()
+        self.component_manager.cleanup()
         super().delete_device()
 
     def _init_state_model(self: MccsFNCC) -> None: