+++ src/ska_low_mccs_pasd/fndh/fndh_component_manager.py
@@ -549,7 +549,7 @@ class FndhComponentManager(TaskExecutorComponentManager):
         result = ResultCode.OK
 
         try:
-            timeout = 60  # seconds
+            timeout = 28 * 10  # seconds
             result, time_left = self._power_fndh_ports(power_state, timeout)
 
         except Exception as ex:  # pylint: disable=broad-except
@@ -614,3 +614,9 @@ class FndhComponentManager(TaskExecutorComponentManager):
         :param value: the value to write.
         """
         self._pasd_bus_proxy.write_attribute(attribute_name, value)
+
+    def cleanup(self: FndhComponentManager) -> None:
+        """Delete and clean up any remaining processes."""
+        if self._pasd_bus_proxy:
+            self._pasd_bus_proxy.cleanup()
+        self._task_executor._executor.shutdown()