+++ src/ska_low_mccs_pasd/pasd_bus/pasd_bus_poll_management.py
@@ -9,6 +9,8 @@
 
 
 import logging
+import time
+from dataclasses import dataclass
 from typing import Any, Callable, Iterator, Optional, Sequence
 
 from ska_low_mccs_pasd.pasd_data import PasdData
@@ -70,7 +72,26 @@ def smartbox_read_request_iterator() -> Iterator[str]:
         yield "ALARM_FLAGS"
 
 
-class DeviceRequestProvider:  # pylint: disable=too-many-instance-attributes
+@dataclass
+class DelayedRequest:
+    """
+    Class to represent a delayed request.
+
+    Encapsulates data about a delayed request including a
+    'not before' timestamp before which the request should
+    not be actioned.
+    """
+
+    device_id: int
+    request_description: tuple[str, Any]
+    not_before: float
+
+    def __post_init__(self) -> None:
+        """Post init method to record the request timestamp."""
+        self.timestamp = time.time()
+
+
+class DeviceRequestProvider:
     """
     A class that determines the next communication with a specified device.
 
@@ -79,10 +100,14 @@ class DeviceRequestProvider:  # pylint: disable=too-many-instance-attributes
     and it decides what, if any, communication with it should occur in the next poll.
     """
 
+    # pylint: disable=too-many-arguments,too-many-instance-attributes
+    # pylint: disable=too-many-positional-arguments
     def __init__(
         self,
         number_of_ports: int,
         read_request_iterator_factory: Callable[[], Iterator[str]],
+        attribute_read_delay: float,
+        port_status_read_delay: float,
         logger: logging.Logger,
     ) -> None:
         """
@@ -91,6 +116,10 @@ class DeviceRequestProvider:  # pylint: disable=too-many-instance-attributes
         :param number_of_ports: the number of ports this device has.
         :param read_request_iterator_factory: a callable that returns
             a read request iterator
+        :param attribute_read_delay: time in seconds to wait after writing an
+            attribute before reading it again
+        :param port_status_read_delay: time in seconds to wait after setting
+            port status before reading it again
         :param logger: a logger.
         """
         self._logger = logger
@@ -113,6 +142,9 @@ class DeviceRequestProvider:  # pylint: disable=too-many-instance-attributes
         # following a write command
         self._attribute_update_requests: list[str] = []
         self._ports_status_update_request: bool = False
+
+        self._attribute_read_delay = attribute_read_delay
+        self._port_status_read_delay = port_status_read_delay
 
         self._read_request_iterator_factory = read_request_iterator_factory
         self._read_request_iterator = read_request_iterator_factory()
@@ -243,9 +275,7 @@ class DeviceRequestProvider:  # pylint: disable=too-many-instance-attributes
                 return "BREAKER_RESET", port
 
         if any(change is not None for change in self._port_power_changes):
-            requested_powers = self._port_power_changes
-            self._port_power_changes = [None] * len(requested_powers)
-            self._ports_status_update_request = True
+            requested_powers = self._get_requested_port_powers()
             return "SET_PORT_POWERS", requested_powers
 
         if self._status_reset_requested:
@@ -264,6 +294,12 @@ class DeviceRequestProvider:  # pylint: disable=too-many-instance-attributes
 
         return "NONE", None
 
+    def _get_requested_port_powers(self) -> list[tuple[bool, bool] | None]:
+        requested_powers = self._port_power_changes
+        self._port_power_changes = [None] * len(requested_powers)
+        self._ports_status_update_request = True
+        return requested_powers
+
     def get_read(self) -> str:
         """
         Return a description of the next read to be performed on the device.
@@ -272,23 +308,136 @@ class DeviceRequestProvider:  # pylint: disable=too-many-instance-attributes
         """
         return next(self._read_request_iterator)
 
-    def get_expedited_read(self) -> tuple[str, Any]:
+    def get_expedited_read(self, device_id: int) -> DelayedRequest | None:
         """
-        Return a description of an expedited read request.
+        Return a DelayedRequest for future action to expedite a read.
 
         This is required for attributes which have been written to by the user,
-        so we don't have to wait until their turn in the regular poll.
+        so we don't have to wait until their turn in the regular poll. The
+        request is delayed to ensure the register has been updated before
+        we read it again.
 
-        :return: A tuple, consisting of the name of a predefined attribute set
-            (see PasdBusComponentManager), or the command READ along with the
-            name of the specific attribute to be read.
+        :param: device_id: The id of the device requiring the request.
+        :return: A DelayedRequest, encapsulating information about
+            the request to make and when it should be actioned.
         """
         if self._ports_status_update_request:
             self._ports_status_update_request = False
-            return "PORTS", None
+            return DelayedRequest(
+                device_id, ("PORTS", None), time.time() + self._port_status_read_delay
+            )
         if self._attribute_update_requests:
-            return "READ", self._attribute_update_requests.pop(0)
-        return "NONE", None
+            return DelayedRequest(
+                device_id,
+                ("READ", self._attribute_update_requests.pop(0)),
+                time.time() + self._attribute_read_delay,
+            )
+        return None
+
+    def get_write_read_sequence(self, device_id: int) -> list[DelayedRequest] | None:
+        """
+        Return a list of DelayedRequests to perform a write-read sequence.
+
+        By default there are none.
+
+        :param device_id: The id of the device requiring the request.
+
+        :returns: a list of DelayedRequests to be serviced.
+        """
+        return None
+
+
+# pylint: disable=too-many-positional-arguments, too-many-arguments
+class FndhRequestProvider(DeviceRequestProvider):
+    """
+    A class to handle staggered powering of Fndh ports.
+
+    A request for powering N ports becomes N requests for powering 1 port.
+    """
+
+    def __init__(
+        self,
+        number_of_ports: int,
+        read_request_iterator_factory: Callable[[], Iterator[str]],
+        attribute_read_delay: float,
+        port_status_read_delay: float,
+        port_power_delay: float,
+        logger: logging.Logger,
+    ) -> None:
+        """
+        Initialise a new instance.
+
+        :param number_of_ports: the number of ports this device has.
+        :param read_request_iterator_factory: a callable that returns
+            a read request iterator
+        :param attribute_read_delay: time in seconds to wait after writing an
+            attribute before reading it again
+        :param port_status_read_delay: time in seconds to wait after setting
+            port status before reading it again
+        :param port_power_delay: time in seconds to wait between setting
+            each FNDH port power.
+        :param logger: a logger.
+        """
+        self._port_power_delay = port_power_delay
+        super().__init__(
+            number_of_ports,
+            read_request_iterator_factory,
+            attribute_read_delay,
+            port_status_read_delay,
+            logger,
+        )
+
+    def _get_requested_port_powers(self) -> list[tuple[bool, bool] | None]:
+        """
+        Get requested FNDH port powers.
+
+        The initial request is simply recorded to be fulfilled through expedited
+        requests later.
+
+        :returns: no ports to change immediately.
+        """
+        return [None] * len(self._port_power_changes)
+
+    def get_write_read_sequence(self, device_id: int) -> list[DelayedRequest] | None:
+        """
+        Get a sequence of write-read requests.
+
+        A command to set port powers will be converted into N requests to write,
+        interleaved with N requests to read.
+
+        :param device_id: The id of the device requiring the request.
+
+        :returns: a list of DelayedRequests.
+        """
+        write_read_sequence = []
+        offset = 0
+        for requested_port, port_power in enumerate(self._port_power_changes):
+            if port_power is not None:
+                requested_powers: list[tuple[bool, bool] | None] = [None] * len(
+                    self._port_power_changes
+                )
+                requested_powers[requested_port] = port_power
+                self._port_power_changes[requested_port] = None
+                port_power_time = time.time() + offset * self._port_power_delay
+                port_read_time = port_power_time + self._port_status_read_delay
+                write_read_sequence.append(
+                    DelayedRequest(
+                        device_id,
+                        ("SET_PORT_POWERS", requested_powers),
+                        port_power_time,
+                    )
+                )
+                write_read_sequence.append(
+                    DelayedRequest(
+                        device_id,
+                        ("PORTS", None),
+                        port_read_time,
+                    )
+                )
+                offset += 1
+        if write_read_sequence:
+            return write_read_sequence
+        return super().get_write_read_sequence(device_id)
 
 
 class PasdBusRequestProvider:
@@ -306,10 +455,15 @@ class PasdBusRequestProvider:
       given the above constraints
     """
 
+    # pylint: disable=too-many-instance-attributes, too-many-arguments
+    # pylint: disable=too-many-positional-arguments
     def __init__(
         self,
         min_ticks: int,
         logger: logging.Logger,
+        attribute_read_delay: float,
+        port_status_read_delay: float,
+        port_power_delay: float,
         available_smartboxes: list[int],
         smartbox_ids: Optional[list[int]] = None,
     ) -> None:
@@ -319,12 +473,28 @@ class PasdBusRequestProvider:
         :param min_ticks: minimum number of ticks between communications
             with any given device
         :param logger: a logger.
+        :param attribute_read_delay: time in seconds to wait after writing an
+            attribute before reading it again
+        :param port_status_read_delay: time in seconds to wait after setting
+            port status before reading it again
+        :param port_power_delay: time in seconds to wait between setting
+            each FNDH port power.
         :param available_smartboxes: list of available smartbox ids to poll
         :param smartbox_ids: optional list of smartbox IDs associated with
             each FNDH port
         """
+        if port_status_read_delay >= port_power_delay:
+            logger.warning(
+                f"Port status read delay ({port_status_read_delay}) >= Port Power Delay"
+                f" ({port_power_delay}), setting port status read delay to "
+                f"{port_power_delay - 1}"
+            )
+            port_status_read_delay = port_power_delay - 1
         self._min_ticks = min_ticks
         self._logger = logger
+        self._attribute_read_delay = attribute_read_delay
+        self._port_status_read_delay = port_status_read_delay
+        self._port_power_delay = port_power_delay
 
         # Create a dict mapping FNDH ports to smartbox Modbus IDs
         # if smartboxIDs is provided, otherwise just use available_smartboxes
@@ -337,6 +507,7 @@ class PasdBusRequestProvider:
         else:
             self._available_smartboxes = available_smartboxes
 
+        self._delayed_requests: list[DelayedRequest] = []
         self.initialise()
 
     def initialise(self) -> None:
@@ -363,16 +534,27 @@ class PasdBusRequestProvider:
                 }
             )
 
-        fndh_request_provider = DeviceRequestProvider(
-            PasdData.NUMBER_OF_FNDH_PORTS, fndh_read_request_iterator, self._logger
+        fndh_request_provider = FndhRequestProvider(
+            PasdData.NUMBER_OF_FNDH_PORTS,
+            fndh_read_request_iterator,
+            self._attribute_read_delay,
+            self._port_status_read_delay,
+            self._port_power_delay,
+            logger=self._logger,
         )
         fncc_request_provider = DeviceRequestProvider(
-            0, fncc_read_request_iterator, self._logger
+            0,
+            fncc_read_request_iterator,
+            attribute_read_delay=self._attribute_read_delay,
+            port_status_read_delay=self._port_status_read_delay,
+            logger=self._logger,
         )
         self._device_request_providers: dict[int, DeviceRequestProvider] = {
             smartbox_id: DeviceRequestProvider(
                 PasdData.NUMBER_OF_SMARTBOX_PORTS,
                 smartbox_read_request_iterator,
+                self._attribute_read_delay,
+                self._port_status_read_delay,
                 self._logger,
             )
             for smartbox_id in self._available_smartboxes
@@ -402,12 +584,29 @@ class PasdBusRequestProvider:
                 self._logger.info(f"Stopping polling smartbox {smartbox_id}")
                 self._ticks.pop(smartbox_id, None)
 
+    def stop_polling_smartboxes(
+        self, port_power_requests: list[tuple[bool, bool] | None]
+    ) -> None:
+        """
+        Stop polling a smartbox if it has been requested to switch off.
+
+        :param port_power_requests: list of port power requests
+        """
+        for fndh_port, request in enumerate(port_power_requests, start=1):
+            if request is not None and request[0] is False:
+                smartbox_id = self._smartboxIDs.get(fndh_port)
+                if smartbox_id is not None and smartbox_id in self._ticks:
+                    self._logger.info(
+                        f"Stopping polling smartbox {smartbox_id} as port "
+                        f"{fndh_port} is being powered off"
+                    )
+                    self._ticks.pop(smartbox_id, None)
+
     def desire_read_startup_info(self, device_id: int) -> None:
         """
         Register a request to read the information usually just read at startup.
 
         :param device_id: the device number.
-            This is 0 for the FNDH and 100 for the FNCC, otherwise a smartbox number.
         """
         self._device_request_providers[device_id].desire_read_startup_info()
 
@@ -416,7 +615,6 @@ class PasdBusRequestProvider:
         Register a request to initialize a device.
 
         :param device_id: the device number.
-            This is 0 for the FNDH, otherwise a smartbox number.
         """
         self._device_request_providers[device_id].desire_initialize()
 
@@ -427,7 +625,6 @@ class PasdBusRequestProvider:
         Register a request to write an attribute.
 
         :param device_id: the device number.
-            This is 0 for the FNDH, otherwise a smartbox number.
         :param name: the name of the attribute to write.
         :param values: the new value(s) to write.
         """
@@ -438,7 +635,6 @@ class PasdBusRequestProvider:
         Register a request to reset an alarm.
 
         :param device_id: the device number.
-            This is 0 for the FNDH, otherwise a smartbox number.
         """
         self._device_request_providers[device_id].desire_alarm_reset()
 
@@ -447,7 +643,6 @@ class PasdBusRequestProvider:
         Register a request to reset a warning.
 
         :param device_id: the device number.
-            This is 0 for the FNDH, otherwise a smartbox number.
         """
         self._device_request_providers[device_id].desire_warning_reset()
 
@@ -477,7 +672,6 @@ class PasdBusRequestProvider:
         Register a request to turn some of device's ports on.
 
         :param device_id: the device number.
-            This is 0 for the FNDH, otherwise a smartbox number.
         :param port_powers: a desired port power state for each port.
             True means the port is desired on,
             False means it is desired off,
@@ -494,7 +688,6 @@ class PasdBusRequestProvider:
         Register a request to reset a port breaker.
 
         :param device_id: the device number.
-            This is 0 for the FNDH, otherwise a smartbox number.
         :param port_number: the number of the port whose breaker is to
             be reset.
         """
@@ -505,7 +698,6 @@ class PasdBusRequestProvider:
         Register a request to set a device's LED pattern.
 
         :param device_id: the device number.
-            This is 0 for the FNDH, otherwise a smartbox number.
         :param pattern: name of the service LED pattern.
         """
         self._device_request_providers[device_id].desire_led_pattern(pattern)
@@ -517,7 +709,6 @@ class PasdBusRequestProvider:
         Register a request to set a device's low pass filter constants.
 
         :param device_id: the device number.
-            This is 0 for the FNDH, otherwise a smartbox number.
         :param cutoff: frequency of LPF to set.
         :param extra_sensors: write the constant to the extra sensors' registers after
             the LED status register.
@@ -564,6 +755,31 @@ class PasdBusRequestProvider:
             self._ticks[PasdData.FNDH_DEVICE_ID] = 0
             return PasdData.FNDH_DEVICE_ID, *("READ", "status")
 
+        # Check if any expedited attribute reads or write-read sequences
+        # need to be added to the list for future polls. These are actioned
+        # after a delay, so we maintain a list rather than executing them
+        # immediately, so as not to hold up other requests.
+        for device_id, _ in self._ticks.items():
+            expedited_read_request = self._device_request_providers[
+                device_id
+            ].get_expedited_read(device_id)
+            if expedited_read_request is not None:
+                self._delayed_requests.append(expedited_read_request)
+            write_read_sequence = self._device_request_providers[
+                device_id
+            ].get_write_read_sequence(device_id)
+            if write_read_sequence is not None:
+                self._delayed_requests.extend(write_read_sequence)
+
+        # Now see if any delayed requests are ready to be executed.
+        # These takes priority over writes so that we can update the polling
+        # list if the port power states have changed.
+        for delayed_request in self._delayed_requests:
+            if delayed_request.not_before < time.time():
+                self._delayed_requests.remove(delayed_request)
+                return delayed_request.device_id, *delayed_request.request_description
+
+        # Next we check for any write requests.
         for device_id, tick in self._ticks.items():
             if tick < self._min_ticks:
                 break
@@ -572,18 +788,6 @@ class PasdBusRequestProvider:
                 del self._ticks[device_id]
                 self._ticks[device_id] = 0
                 return device_id, *write_request
-
-        # Next check if any expedited attribute reads need to be done.
-        for device_id, tick in self._ticks.items():
-            if tick < self._min_ticks:
-                break
-            expedited_read_request = self._device_request_providers[
-                device_id
-            ].get_expedited_read()
-            if expedited_read_request != ("NONE", None):
-                del self._ticks[device_id]  # see comment above
-                self._ticks[device_id] = 0
-                return device_id, *expedited_read_request
 
         # No outstanding reads/writes remaining, so cycle through the polling list.
         fncc_skip = False