+++ CHANGELOG.md
@@ -2,8 +2,26 @@
 
 ## Unreleased
 
-## 4.2.1
+* [THORN-338] SetFndhPortPowers is now executed in a staggered manner. Each port is powered one at a time, with the time between dictated by the PortPowerDelay property on MccsPasdBus. This delay must be greater than the PortStatusReadDelay value, such that we can always fit a read in between powering each port. If configured with PortStatusReadDelay >= PortPowerDelay, the value used will be calculated as PortStatusReadDelay = PortPowerDelay - 1.
+* [THORN-403] Attempt to reconnect after a catastrophic error such as a 'Broken pipe'.
+* [THORN-402] Configure pipeline to ignore sonar scan results.
 
+## 4.3.2
+
+* [SKB-1207] Add a FaultOnThresholdDifference Tango property (default True) to MccsSmartBox and MccsFndh, to control whether the device will enter DevState.FAULT when there is a mismatch between firmware and Tango thresholds.
+* [LOW-2058] Fix typo in attribute name
+* [THORN-383] Fix KeyError caused when not deploying full set of smartboxes
+
+## 4.3.1
+
+* [THORN-399] Include device and attribute name when logging error events
+* [THORN-384] Small fixes for thread cleanup
+
+## 4.3.0
+
+* [THORN-391] Delay next poll after changing FNDH port status. Introduces new properties AttributeReadDelay and PortStatusReadDelay. Also addresses a bugs in switching adminMode.
+
+## 4.2.1
 * [THORN-388] Update to ska-low-pasd-driver 1.0.0 which moves us to pymodbus 3.11.4. 
 * [THORN-336] Request FNPC SYS_STATUS after a Modbus error to try to re-establish communications.
 * [LOW-2034] Only deploy a PaSD configuration ConfigMap if deploying the corresponding simulator that needs it.